+++ 26.1/src/main/kotlin/com/bleudev/nine_lifes/custom/NineLifesCriterions.kt
@@ -11,10 +11,9 @@ object NineLifesCriterions {
     val SUCCESS_SLEEP_WITH_AMETHYSM = create("success_sleep_with_amethysm", SuccessSleepWithAmethysmCriterion())
     val ALMOST_DEAD = create("almost_dead", AlmostDeadCriterion())
     val LIFES_PLAY_TIME = create("lifes_play_time", LifesPlayTimeCriterion())
-    val PLAY_TIME = create("play_time", PlayTimeCriterion())
+    val CUSTOM_STAT = create("custom_stat", CustomStatCriterion())
     val CHARGE_ITEM = create("charge_item", ChargeItemCriterion())
     val LIFES_CHANGE = create("lifes_change", LifesChangeCriterion())
-    val USED_CHARGED_TOTAL = create("used_charged_total", UsedChargedTotalCriterion())
     val ADVANCEMENT = create("advancement", AdvancementCriterion())
 
     private fun <T : CriterionTrigger<*>> create(name: String, criterion: T) = CriteriaTriggers.register(createIdentifier(name).toString(), criterion)
@@ -24,7 +23,6 @@ object NineLifesCriterions {
     internal fun trigger(player: ServerPlayer) {
         ALMOST_DEAD.trigger(player)
         LIFES_PLAY_TIME.trigger(player)
-        PLAY_TIME.trigger(player)
         ADVANCEMENT.trigger(player)
     }
 }
